+++ nodes/hikvisionUltimatePicture.js
@@ -7,6 +7,7 @@ module.exports = function (RED) {
 	function hikvisionUltimatePicture(config) {
 		RED.nodes.createNode(this, config);
 		var node = this;
+		node.topic = config.topic || config.name;
 		node.server = RED.nodes.getNode(config.server)
 		node.picture; // Stores the cam image
 		node.pictureLarghezza = 0;
@@ -85,7 +86,7 @@ module.exports = function (RED) {
 						setTimeout(CB, 500);
 
 					}).catch(error => {
-						res.json({ picture: "", width: " !Error getting picture! ", height: " !" + error.message + "! "});
+						res.json({ picture: "", width: " !Error getting picture! ", height: " !" + error.message + "! " });
 					});
 				}
 			} else { res.json({ picture: "", width: 0, height: 0 }); }
@@ -118,6 +119,7 @@ module.exports = function (RED) {
 		// Called from config node, to send output to the flow
 		node.sendPayload = (_msg) => {
 			if (_msg === null || _msg === undefined) return;
+			_msg.topic = node.topic;
 			if (_msg.hasOwnProperty("errorDescription")) { node.send([null, _msg]); return; }; // It's a connection error/restore comunication.
 			if (_msg.hasOwnProperty("payload")) {
 				if (_msg.payload.hasOwnProperty("eventType")) {