+++ nodes/hikvisionUltimateRadar.js
@@ -4,18 +4,62 @@ module.exports = function (RED) {
 	function hikvisionUltimateRadar(config) {
 		RED.nodes.createNode(this, config);
 		var node = this;
+		node.topic = config.topic || config.name;
 		node.server = RED.nodes.getNode(config.server)
 		node.filterzone = config.filterzone || "0";// Rect to alarm coming from...
+		node.alarmfilterduration = config.alarmfilterduration !== undefined ? config.alarmfilterduration : 0;
+		node.alarmfilterperiod = config.alarmfilterperiod !== undefined ? config.alarmfilterperiod : 0;
+		node.currentAlarmMSG = {}; // Stores the current alarm object
+		node.total_alarmfilterduration = 0; // stores the total time an alarm has been true in the alarmfilterperiod time.
+		node.isNodeInAlarm = false; // Stores the current state of the filtered alarm.
+		node.isRunningTimerFilterPeriod = false; // Indicates wether the period timer is running;
 
 		node.setNodeStatus = ({ fill, shape, text }) => {
 			var dDate = new Date();
 			node.status({ fill: fill, shape: shape, text: text + " (" + dDate.getDate() + ", " + dDate.toLocaleTimeString() + ")" })
 		}
 
+		// 15/01/2021 start the timer that counts the time the alarm has been true
+		// ###################################
+		node.timer_alarmfilterduration = setInterval(() => {
+			if (node.currentAlarmMSG.hasOwnProperty("payload")) {
+				if (node.currentAlarmMSG.payload === true) {
+					node.total_alarmfilterduration += 1;
+					if (node.isRunningTimerFilterPeriod) node.setNodeStatus({ fill: "red", shape: "ring", text: "Zone " + node.currentAlarmMSG.zone + " pre alert count " + node.total_alarmfilterduration });
+					if (node.total_alarmfilterduration >= node.alarmfilterduration) {
+						if (!node.isNodeInAlarm) {
+							// Emit alarm
+							if (node.timer_alarmfilterperiod !== null) clearTimeout(node.timer_alarmfilterperiod);
+							//node.setNodeStatus({ fill: "yellow", shape: "ring", text: "STOP TimerFilterPeriod" });
+							node.isRunningTimerFilterPeriod = false;
+							node.isNodeInAlarm = true;
+							node.total_alarmfilterduration = 0;
+							node.setNodeStatus({ fill: "red", shape: "dot", text: "Zone " + node.currentAlarmMSG.zone + " alarm" });
+							node.send([node.currentAlarmMSG, null]);
+						} else { node.total_alarmfilterduration = 0; }
+					}
+				}
+			}
+		}, 1000);
+		// This timer resets the node.total_alarmfilterduration
+		startTimerFilterPeriod = () => {
+			//node.setNodeStatus({ fill: "yellow", shape: "ring", text: "START TimerFilterPeriod" });
+			node.isRunningTimerFilterPeriod = true;
+			node.total_alarmfilterduration = 0;
+			node.timer_alarmfilterperiod = setTimeout(() => {
+				node.total_alarmfilterduration = 0;
+				//node.setNodeStatus({ fill: "yellow", shape: "ring", text: "ELAPSED TimerFilterPeriod" });
+				node.isRunningTimerFilterPeriod = false;
+			}, node.alarmfilterperiod * 1000);
+		}
+		// ###################################
+
+
 		// Called from config node, to send output to the flow
 		node.sendPayload = (_msg) => {
 			if (_msg === null || _msg === undefined) return;
-			if (_msg.hasOwnProperty("errorDescription")) { node.send([null,_msg]); return; }; // It's a connection error/restore comunication.
+			_msg.topic = node.topic;
+			if (_msg.hasOwnProperty("errorDescription")) { node.send([null, _msg]); return; }; // It's a connection error/restore comunication.
 			// Radar alarm JSON
 			/* 	{
 					"ipAddress": "192.168.1.25",
@@ -48,12 +92,13 @@ module.exports = function (RED) {
 					}
 				} */
 			//var oRetMsg = RED.util.cloneMessage(_msg);
-		
+
 			if (_msg.hasOwnProperty("payload")
 				&& _msg.payload.hasOwnProperty("CIDEvent")
 				&& _msg.payload.CIDEvent.type.toString().toLowerCase() === "zonealarm"
 				&& _msg.payload.CIDEvent.hasOwnProperty("zone")) {
-				var oRetMsg = {}; // Return message
+				let oRetMsg = {}; // Return message
+				oRetMsg.topic = _msg.topic;
 				oRetMsg.connected = _msg.connected;
 				oRetMsg.alarm = _msg.payload; // Put the full alarm description here.
 				oRetMsg.zone = _msg.payload.CIDEvent.zone + 1; // The zone on device's ISAPI is base 0, while the zone on UI is base 1.
@@ -63,7 +108,7 @@ module.exports = function (RED) {
 						case "1103":
 							// Start alarm. Standard SIA Code is _msg.payload.CIDEvent.standardCIDcode: 1130
 							oRetMsg.payload = true;
-							node.setNodeStatus({ fill: "red", shape: "dot", text: "Zone " + oRetMsg.zone + " alert" });
+							node.setNodeStatus({ fill: "red", shape: "ring", text: "Zone " + oRetMsg.zone + "  pre alert" });
 							break;
 						case "3103":
 							// End alarm. Standard SIA Code is _msg.payload.CIDEvent.standardCIDcode: 3130
@@ -76,7 +121,17 @@ module.exports = function (RED) {
 							node.setNodeStatus({ fill: "grey", shape: "ring", text: "Zone " + oRetMsg.zone + " unknowk CID code " + _msg.payload.CIDEvent.code });
 							return; // Unknown state
 					}
-					node.send([oRetMsg,null]);
+					node.currentAlarmMSG = oRetMsg;
+					// Sends the false only in case the isNodeInAlarm is true.
+					if (oRetMsg.payload === false && node.isNodeInAlarm) {
+						node.send([oRetMsg, null]);
+						node.currentAlarmMSG = {};
+						node.isNodeInAlarm = false;
+					} else if (oRetMsg.payload === true && !node.isNodeInAlarm) {
+						if (!node.isRunningTimerFilterPeriod) {
+							startTimerFilterPeriod();
+						}
+					}
 				}
 			}
 		}
@@ -88,10 +143,12 @@ module.exports = function (RED) {
 		}
 
 		this.on('input', function (msg) {
-		
+			node.sendPayload(msg);
 		});
 
 		node.on("close", function (done) {
+			if (node.timer_alarmfilterduration !== null) clearInterval(node.timer_alarmfilterduration);
+			if (node.timer_alarmfilterperiod !== null) clearTimeout(node.timer_alarmfilterperiod);
 			if (node.server) {
 				node.server.removeClient(node);
 			}